Integer --> Byte valtozok

+++ RGB_img/Models/Pixel.cs
@@ -8,11 +8,11 @@ namespace RGB_img.Models
 {
     internal class Pixel
     {
-        public int R { get; set; }
-        public int G { get; set; }
-        public int B { get; set; }
+        public byte R { get; set; }
+        public byte G { get; set; }
+        public byte B { get; set; }
 
-        public Pixel(int r, int g, int b)
+        public Pixel(byte r, byte g, byte b)
         {
             R = r;
             G = g;
